+++ backend/infrastructures/config/settings/injections_configurations.py
@@ -1,7 +1,9 @@
 import inject
 
+from ...apps.extract.adapters import units_of_work as extract_unit_of_work
 from ...apps.load.adapters import repositories as query_load_repositories
 from ...apps.load.adapters import units_of_work as load_unit_of_work
+
 
 
 def inject_config(binder: inject.Binder):
@@ -12,7 +14,7 @@ def inject_config(binder: inject.Binder):
         "query_data_repository", query_load_repositories.DjangoDataQueryRepository
     )
     binder.bind_to_constructor(
-        "save_file_repository", query_load_repositories.DjangoFileSaveRepository
+        "file_unit_of_work", extract_unit_of_work.DjangoFileUnitOfWork
     )
