+++ backend/modules/load/domain/ports/repositories.py
@@ -1,6 +1,4 @@
-import io
 from abc import abstractmethod
-from pathlib import Path
 
 from ....common.domain import ports as common_ports
 from ...services import queries as load_queries
@@ -12,14 +10,5 @@ class AbstractDataDomainRepository(common_ports.AbstractDomainRepository):
         """
         :param: Data to save.
         :raises DataCreationException.
-        """
-        pass
-
-class AbstractFileSaveRepository(common_ports.AbstractDomainRepository):
-    @abstractmethod
-    def save(self, bytes: io.BytesIO, file_name: str) -> Path:
-        """
-        :param: File as bytes.
-        :raises Path where file is saved.
         """
         pass
